[➕Update] Refactor code
[Update 메서드 호출 수 감소]
- 밭과 자원 오브젝트의 Update를 DayCycle 이벤트로 이관

[플레이어 콤보어택 버그 수정]
- 광클 시 전진하던 문제 수정
- 광클 시 다음 공격이 나가지 않던 문제 수정

[몬스터 스폰 무한루프 문제 수정]

[카메라를 가리는 오브젝트 관련]
- Trigger 콜라이더도 접촉 가능하게 수정
- 나무 프리팹 콜라이더 변경

+++ Assets/02. Scripts/Data/ArchitectureSaveData.cs
@@ -43,7 +43,7 @@ public struct FarmData
     }
 
     public int state;
-    public float remainingTime;
+    public int remainingTime;
 }
 
 [System.Serializable]
