+++ Assets/02. Scripts/Resource/ResourceObjectDebris.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 // 2024-01-24 WJY
 public class ResourceObjectDebris : ResourceObjectBase
 {
-    [SerializeField] private float _respawnTime;
-    public float RespawnTime => _respawnTime;
+    [SerializeField] private int _respawnTime;
+    public int RespawnTime => _respawnTime;
 
 
     public void Respawn()
